% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{characters}
\alias{characters}
\title{Characters.}
\format{A data frame with four variables:
\describe{
\item{\code{id}}{an ID assigned to each character}
\item{\code{name}}{character name}
\item{\code{normalized_name}}{character normalized name}
\item{\code{sex}}{character sex}
}}
\usage{
characters
}
\description{
Contains The Simpsons characters information.
}
\keyword{datasets}
