% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{episodes}
\alias{episodes}
\title{Episodes.}
\format{A data frame with thirteen variables:
\describe{
\item{\code{id}}{an ID assigned to each episode}
\item{\code{title}}{episode title}
\item{\code{original_air_date}}{original air date}
\item{\code{production_code}}{production code}
\item{\code{season}}{episode's season}
\item{\code{number_in_season}}{episode number inside the season}
\item{\code{number_in_series}}{episode number in the serie}
\item{\code{us_viewers_in_millions}}{US viewers in millions}
\item{\code{views}}{views}
\item{\code{imdb_rating}}{IMDb episode rating}
\item{\code{imdb_votes}}{IMDb episode votes}
\item{\code{image_url}}{episode image url}
\item{\code{video_url}}{episode video url}
}}
\usage{
episodes
}
\description{
Contains metadata related to each The Simpsons episode.
}
\keyword{datasets}
