% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations.}
\format{A data frame with three variables:
\describe{
\item{\code{id}}{an ID assigned to each location}
\item{\code{name}}{location name}
\item{\code{normalized_name}}{location normalized name}
}}
\usage{
locations
}
\description{
Contains The Simpsons world locations.
}
\keyword{datasets}
