% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{script_lines}
\alias{script_lines}
\title{Script Lines.}
\format{A data frame with thirteen variables:
\describe{
\item{\code{id}}{an ID assigned to each script line}
\item{\code{episode_id}}{associated episode ID in which this line happens}
\item{\code{number}}{line number in each episode}
\item{\code{raw_text}}{raw text}
\item{\code{timestamp_in_ms}}{line timestap}
\item{\code{speaking_line}}{if the line was spoken}
\item{\code{character_id}}{associated character ID that speaks}
\item{\code{raw_character_text}}{speaking character raw name}
\item{\code{location_id}}{associated location ID in which this line happens}
\item{\code{raw_location_text}}{location raw name}
\item{\code{spoken_words}}{text}
\item{\code{normalized_text}}{normalized text}
\item{\code{word_count}}{line number of words}
}}
\usage{
script_lines
}
\description{
Contains the text spoken during each The Simpsons episode (including details
about which character said it and where).
}
\keyword{datasets}
